Correct composer install, stop warnings about file existings, correct build file

+++ silverstripe-web-container/7.1/resources/scripts/entrypoint.sh
@@ -36,14 +36,6 @@ fi
 mkdir /var/www/html/public/assets
 chown -R www-data:www-data /var/www/html/public/assets
 
-
-#
-# Change permissions on .profile folder
-#
-mkdir /var/www/html/.profile
-chown -R www-data:www-data /var/www/html/.profile
-
-
 #
 # Create needs-setup file to trigger the setup-page
 #
@@ -60,7 +52,9 @@ chmod 777 /cache
 #
 # Add symbolic link to wkhtml2pdf
 #
-ln -s /usr/local/bin/wkhtmltopdf /usr/local/bin/wkhtmltopdf_12
+if [ ! -f "/usr/local/bin/wkhtmltopdf_12" ]; then
+	ln -s /usr/local/bin/wkhtmltopdf /usr/local/bin/wkhtmltopdf_12
+fi
 
 
 exec gosu $HOST_UID:$HOST_GID /fork.sh
